Use dict instead of orddict in erl_lint variable table
Once you have more than a dozen of variables in the same
context, the performance with orddict really starts to
deteriorate. Using dict gives better performance overall.

Furthermore, this patch makes the code uniform in using
vtnew() instead of directly creating new variable tables
using []. This means it should be relatively straight-forward
to migrate to another data-structure if desired in the
future.