Use full list of bindings when matching on map keys
Prior to this patch, the following code would not eval:

    X = key,
    (fun(#{X := value}) -> true end)(#{X => value})

That's because the key evaluation was using the new
list of bindings introduced on every anonymous fun.
Since keys only match on values defined prior to the
pattern and do not introduce any new binding, we must
use the full and original list of binding.