Add code path caching
When an application has several entries in its code path,
loading modules in interactive mode becomes more expensive
because of code path misses.

This commit introduces code path caching, where we can opt-in
into caching each directory individually.

This won’t be applied to all paths but a project has several paths
that are unlikely to change while a system is running in development
or test:

  * the paths from Erlang/OTP won’t be written to by most projects
  * paths from build tools and other languages
  * non-local dependencies, such as the ones from Hex/Git

TODO
----

  * [ ] Currently this commit enables caching by defaut.
    In practice, we want to introduce several control mechanisms.

  * [ ] Docs!